+++ apps/trackers/tests/test_common.py
@@ -5,7 +5,7 @@ import pytest
 from django.utils import timezone
 
 from apps.trackers.common import TrackerQueryBuilder
-from osidb.models import Affect, Flaw, Tracker
+from osidb.models import Affect, ContractPriority, Flaw, Tracker
 from osidb.tests.factories import (
     AffectFactory,
     FlawFactory,
@@ -327,10 +327,9 @@ class TestTrackerQueryBuilderDescription:
         tqb.instance = tracker
 
         assert tqb.description == (
-            """\
-special-module tracking bug for large-component: see the bugs linked in the "Blocks" field of this bug for full details of the security issue(s).
-
-This bug is never intended to be made public, please put any public notes in the blocked bugs."""
+            "special-module tracking bug for large-component: see the bugs linked "
+            'in the "Blocks" field of this bug for full details of the security issue(s).\n\n'
+            "This bug is never intended to be made public, please put any public notes in the blocked bugs."
         )
 
     def test_bugzilla_embargoed(self):
@@ -371,14 +370,11 @@ This bug is never intended to be made public, please put any public notes in the
         tqb.instance = tracker
 
         assert tqb.description == (
-            """\
-special-module tracking bug for large-component: see the bugs linked in the "Blocks" field of this bug for full details of the security issue(s).
-
-This bug is never intended to be made public, please put any public notes in the blocked bugs.
-
-NOTE THIS ISSUE IS CURRENTLY EMBARGOED, DO NOT MAKE PUBLIC COMMITS OR COMMENTS ABOUT THIS ISSUE.
-
-WARNING: NOTICE THAT REMOVING THE "SECURITY" GROUP FROM THIS TRACKER MAY BREAK THE EMBARGO."""
+            "special-module tracking bug for large-component: see the bugs linked "
+            'in the "Blocks" field of this bug for full details of the security issue(s).\n\n'
+            "This bug is never intended to be made public, please put any public notes in the blocked bugs.\n\n"
+            "NOTE THIS ISSUE IS CURRENTLY EMBARGOED, DO NOT MAKE PUBLIC COMMITS OR COMMENTS ABOUT THIS ISSUE.\n\n"
+            'WARNING: NOTICE THAT REMOVING THE "SECURITY" GROUP FROM THIS TRACKER MAY BREAK THE EMBARGO.'
         )
 
     def test_bugzilla_rhel(self):
@@ -405,13 +401,9 @@ WARNING: NOTICE THAT REMOVING THE "SECURITY" GROUP FROM THIS TRACKER MAY BREAK T
         tqb.instance = tracker
 
         assert (
-            (
-                """\
-For the Enterprise Linux security issues handling process overview see:
-https://source.redhat.com/groups/public/product-security/content/product_security_wiki/eus_z_stream_and_security_bugs"""
-            )
-            in tqb.description
-        )
+            "For the Enterprise Linux security issues handling process overview see:\n"
+            "https://source.redhat.com/groups/public/product-security/content/product_security_wiki/eus_z_stream_and_security_bugs"
+        ) in tqb.description
 
     @pytest.mark.parametrize("ps_component", ["xen", "kvm", "kernel-xen"])
     def test_bugzilla_virtualizaiton(self, ps_component):
@@ -439,14 +431,53 @@ https://source.redhat.com/groups/public/product-security/content/product_securit
         tqb.instance = tracker
 
         assert (
-            (
-                """\
-Information with regards to this bug is considered Red Hat Confidential \
-until the embargo has lifted. Please post the patch only to the \
-'rhkernel-team-list' and/or 'virt-devel' mailing lists for review and acks."""
-            )
-            in tqb.description
+            "Information with regards to this bug is considered Red Hat Confidential "
+            "until the embargo has lifted. Please post the patch only to the "
+            "'rhkernel-team-list' and/or 'virt-devel' mailing lists for review and acks."
+        ) in tqb.description
+
+    @pytest.mark.parametrize("is_contract_priority", [True, False])
+    def test_contract_priority(self, is_contract_priority):
+        """
+        test contract priority description generation
+        """
+        ps_module = PsModuleFactory(bts_name="bugzilla")
+        ps_update_stream = PsUpdateStreamFactory(ps_module=ps_module)
+        if is_contract_priority:
+            ContractPriority(ps_update_stream=ps_update_stream.name).save()
+        flaw = FlawFactory()
+        affect = AffectFactory(
+            flaw=flaw,
+            affectedness=Affect.AffectAffectedness.AFFECTED,
+            resolution=Affect.AffectResolution.DELEGATED,
+            ps_module=ps_module.name,
+            ps_component="kvm",
         )
+        tracker = TrackerFactory(
+            affects=[affect],
+            embargoed=affect.flaw.embargoed,
+            ps_update_stream=ps_update_stream.name,
+            type=Tracker.TrackerType.BUGZILLA,
+        )
+
+        tqb = TrackerQueryBuilder()
+        tqb.instance = tracker
+
+        description = (
+            "Contract priorities are scoped by at the "
+            "product/version level and include all severities "
+            "and all components within that product/version. "
+            "All severities (including low) must be resolved "
+            "within the allotted SLA time."
+        )
+        faq_url = "https://source.redhat.com/departments/products_and_global_engineering/product_security/ops/product_security_wiki/contract_prioritization_faq"
+
+        if is_contract_priority:
+            assert faq_url in tqb.description
+            assert description in tqb.description
+        else:
+            assert faq_url not in tqb.description
+            assert description not in tqb.description
 
     @pytest.mark.parametrize("flaw_count", [1, 2, 3])
     def test_community(self, flaw_count):
@@ -493,39 +524,32 @@ until the embargo has lifted. Please post the patch only to the \
 
         if flaw_count == 1:
             assert tqb.description == (
-                """\
-More information about this security flaw is available in the following bug:
-
-https://example.com/show_bug.cgi?id=0
-
-Disclaimer: Community trackers are created by Red Hat Product Security team on a \
-best effort basis. Package maintainers are required to ascertain if the flaw indeed \
-affects their package, before starting the update process."""
+                "More information about this security flaw is available in the following bug:\n\n"
+                "https://example.com/show_bug.cgi?id=0\n\n"
+                "Disclaimer: Community trackers are created by Red Hat Product Security team on a "
+                "best effort basis. Package maintainers are required to ascertain if the flaw indeed "
+                "affects their package, before starting the update process."
             )
         elif flaw_count == 2:
             assert tqb.description == (
-                """\
-More information about these security flaws is available in the following bugs:
-
-https://example.com/show_bug.cgi?id=0
-https://example.com/show_bug.cgi?id=1
-
-Disclaimer: Community trackers are created by Red Hat Product Security team on a \
-best effort basis. Package maintainers are required to ascertain if the flaw indeed \
-affects their package, before starting the update process."""
+                "More information about these security flaws is available in the following bugs:\n\n"
+                "https://example.com/show_bug.cgi?id=0\n"
+                "https://example.com/show_bug.cgi?id=1\n\n"
+                "Disclaimer: Community trackers are created by Red Hat Product Security team on a "
+                "best effort basis. Package maintainers are required to ascertain if the flaw indeed "
+                "affects their package, before starting the update process."
             )
         elif flaw_count == 3:
             assert tqb.description == (
-                """\
-More information about these security flaws is available in the following bugs:
-
-https://example.com/show_bug.cgi?id=0
-https://example.com/show_bug.cgi?id=1
-https://example.com/show_bug.cgi?id=2
-
-Disclaimer: Community trackers are created by Red Hat Product Security team on a \
-best effort basis. Package maintainers are required to ascertain if the flaw indeed \
-affects their package, before starting the update process."""
+                (
+                    "More information about these security flaws is available in the following bugs:\n\n"
+                    "https://example.com/show_bug.cgi?id=0\n"
+                    "https://example.com/show_bug.cgi?id=1\n"
+                    "https://example.com/show_bug.cgi?id=2\n\n"
+                    "Disclaimer: Community trackers are created by Red Hat Product Security team on a "
+                    "best effort basis. Package maintainers are required to ascertain if the flaw indeed "
+                    "affects their package, before starting the update process."
+                )
             )
 
     def test_jira_basic(self):
@@ -568,20 +592,14 @@ affects their package, before starting the update process."""
         tqb.instance = tracker
 
         assert tqb.description == (
-            """\
-Security Tracking Issue
-
-Do not make this issue public.
-
-Flaw:
------
-
-serious flaw
-https://example.com/show_bug.cgi?id=12345
-
-this flaw is very hard to fix
-
-~~~"""
+            "Security Tracking Issue\n\n"
+            "Do not make this issue public.\n\n"
+            "Flaw:\n"
+            "-----\n\n"
+            "serious flaw\n"
+            "https://example.com/show_bug.cgi?id=12345\n\n"
+            "this flaw is very hard to fix\n\n"
+            "~~~"
         )
 
     def test_jira_embargoed(self):
@@ -624,24 +642,16 @@ this flaw is very hard to fix
         tqb.instance = tracker
 
         assert tqb.description == (
-            """\
-Security Tracking Issue
-
-Do not make this issue public.
-
-NOTE THIS ISSUE IS CURRENTLY EMBARGOED, DO NOT MAKE PUBLIC COMMITS OR COMMENTS ABOUT THIS ISSUE.
-
-WARNING: NOTICE THAT CHANGING THE SECURITY LEVEL FROM "SECURITY ISSUE" TO "RED HAT INTERNAL" MAY BREAK THE EMBARGO.
-
-Flaw:
------
-
-serious flaw
-https://example.com/show_bug.cgi?id=12345
-
-this flaw is very hard to fix
-
-~~~"""
+            "Security Tracking Issue\n\n"
+            "Do not make this issue public.\n\n"
+            "NOTE THIS ISSUE IS CURRENTLY EMBARGOED, DO NOT MAKE PUBLIC COMMITS OR COMMENTS ABOUT THIS ISSUE.\n\n"
+            'WARNING: NOTICE THAT CHANGING THE SECURITY LEVEL FROM "SECURITY ISSUE" TO "RED HAT INTERNAL" MAY BREAK THE EMBARGO.\n\n'
+            "Flaw:\n"
+            "-----\n\n"
+            "serious flaw\n"
+            "https://example.com/show_bug.cgi?id=12345\n\n"
+            "this flaw is very hard to fix\n\n"
+            "~~~"
         )
 
     @pytest.mark.parametrize("bts_name", ["bugzilla", "jboss"])
@@ -675,14 +685,10 @@ this flaw is very hard to fix
 
         if bts_name == "bugzilla" and embargoed:
             assert (
-                (
-                    """\
-Information with regards to this bug is considered Red Hat Confidential \
-until the embargo has lifted. Please post the patch only to the \
-'rhkernel-team-list' mailing list for review and acks."""
-                )
-                in tqb.description
-            )
+                "Information with regards to this bug is considered Red Hat Confidential "
+                "until the embargo has lifted. Please post the patch only to the "
+                "'rhkernel-team-list' mailing list for review and acks."
+            ) in tqb.description
 
         assert tqb.description.endswith(
             "Reproducers, if any, will remain confidential and never be made public, unless done so by the security team."