+++ docs/CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Duplicate Alerts created concurrently in multiple threads handled correctly (OSIDB-3048)
 - Make task collector ignore outdated issues (OSIDB-3085)
 - Allow Flaw API to proper unassign owner in Jira (OSIDB-3145)
+- Remove sync from Bugzilla from the async sync to Bugzilla (OSIDB-3199)
 
 ## [4.1.3] - 2024-07-25
 ### Changed